+++ app-backend/src/__tests__/utils.test.ts
@@ -26,7 +26,8 @@ jest.mock('aws-sdk', () => {
     };
 });
 
-import {getEnvironmentVariables, addEnvironmentToTableName} from '../utils';
+import {getEnvironmentVariables, addEnvironmentToTableName, hasAuthorizedRoles} from '../utils';
+import {AuthorizerAttributes} from '../interfaces';
 
 const unroll = require('unroll');
 unroll.use(it);
@@ -42,7 +43,7 @@ describe('Test for getEnvironmentVariables', () => {
     unroll('it should #description when #expectedResult', async (
             done: () => void,
             args: {expectedResult: string, description: string}
-    ): Promise<void> => {
+    ) => {
         const result = await getEnvironmentVariables();
         if (args.expectedResult === 'pass') {
             expect(result).toEqual({'demo': 'test'});
@@ -63,5 +64,27 @@ describe('Test for addEnvironmentToTableName', () => {
     it('should add the environment to the table name.', () => {
         const result: string = addEnvironmentToTableName('test-table', '01');
         expect(result).toBe('dev-01-test-table');
+    });
+});
+
+describe('Tests for hasAuthorizedRoles', () => {
+    unroll('it should throw an error when the authorizer is #authorizerValue', async (
+            done: () => void,
+            args: {authorizerValue: AuthorizerAttributes}
+    ) => {
+        expect(() => hasAuthorizedRoles(args.authorizerValue, ['ADMIN'])).toThrowError('The user is unauthorized.');
+        done();
+    }, [
+        ['authorizerValue'],
+        [''],
+        [{claims: ''}],
+        [{claims: {}}],
+        [{claims: {'custom:roles': ''}}],
+        [{claims: {'custom:roles': 'USER'}}],
+    ]);
+
+    it('should return true when the current user is an admin.', () => {
+        const result: boolean = hasAuthorizedRoles({claims: {'custom:roles': 'USER, ADMIN'}}, ['ADMIN']);
+        expect(result).toEqual(true);
     });
 });