+++ app-backend/src/utils.ts
@@ -8,6 +8,7 @@
 
 import * as AWS from 'aws-sdk';
 import * as Bluebird from 'bluebird';
+import {AuthorizerAttributes} from './interfaces';
 
 type DecryptRequest = AWS.KMS.Types.DecryptRequest;
 type DecryptResponse = AWS.KMS.Types.DecryptResponse;
@@ -35,4 +36,27 @@ export async function getEnvironmentVariables(): Promise<Object | void> {
  */
 export function addEnvironmentToTableName(tableName: string, version: string): string {
     return `${process.env.NODE_ENV}-${version}-${tableName}`;
+}
+
+/**
+ * Utility function to check if a user is authorized or not.
+ */
+export function hasAuthorizedRoles(authorizer: AuthorizerAttributes, allowedRoles: string[]): boolean {
+    let isAuthorized: boolean = true;
+
+    if (!authorizer || !authorizer.claims || !authorizer.claims[`custom:roles`]) {
+        throw new Error('The user is unauthorized.');
+    }
+
+    const currentUserRoles: string[] = authorizer.claims[`custom:roles`].split(',');
+
+    isAuthorized = currentUserRoles.some((role: string): boolean => {
+        return (allowedRoles.indexOf(role.trim()) > -1);
+    });
+
+    if (!isAuthorized) {
+        throw new Error('The user is unauthorized.');
+    }
+
+    return isAuthorized;
 }